ml-services/group_auth/requirements.txt
flask==3.0.0
flask-cors==4.0.0
torch==2.1.2
torchvision==0.16.2
opencv-python==4.9.0.80
mtcnn==0.1.1
facenet-pytorch==2.5.3
numpy==1.24.3
pillow==10.1.0
